Change to the way constructors are called
Previously an array was allocated at construction of the empty queue in which every element of the array was default constructed and then reassigned by push().  Now all space is allocated as an array of chars (which have no default constructor) and constructors (and now destructors too) are called explicitly for push() and pop().  This paves the way for implementation of an emplace() function.